requirements.txt
requests==2.29.0
fastapi==0.95.1
uvicorn==0.22.0
python-dotenv==1.0.0
Jinja2==3.1.2